@@ -1,14 +1,15 @@
 import "regenerator-runtime/runtime.js";
 
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 
 import App from "./App";
 import { BrowserRouter } from "react-router-dom";
 
 import "./global.css";
 
-const root = document.querySelector("#root");
+const container = document.querySelector("#root");
+const root = createRoot(container);
 
 const app = (
   <>
@@ -18,4 +19,4 @@ const app = (
   </>
 );
 
-ReactDOM.render(app, root);
+root.render(app);
